+++ src/preload.js
@@ -26,38 +26,6 @@ contextBridge.exposeInMainWorld('file', {
    deleteFile: (filePath) => ipcRenderer.invoke('delete-file', filePath), //delete file
    showMessageBox: (options) => ipcRenderer.invoke('show-message-box', options), //message box to replace alerts
    generatePreviewHTML: async (filepath) => await generatePreviewHTML(filepath),
-   convertDocxToPdf: async (filepath) => {
-      /// Converts from DOCX to PDF for previewing purposes.
-      var html;
-
-      await mammoth.convertToHtml({ path: filepath })
-         .then(function(result) {
-            html = result.value;
-            // Any messages, such as warnings during conversion
-            var _messages = result.messages;
-         })
-         .catch(function(error) {
-            console.error(error);
-            return;
-         });
-
-      // Preliminary pdfMake configuration.
-      pdfMake.vfs = pdfFonts;
-
-      // Create new DOM window object.
-      const { window } = new JSDOM('');
-
-      const converted = htmlToPdfmake(html, { window });
-      const docDefinition = { content: converted };
-
-      pdfPath = path.join(os.tmpdir(), "docxToPdf.pdf");
-
-      pdfMake.createPdf(docDefinition).getBuffer((buffer) => {
-         require('fs').writeFileSync(pdfPath, buffer);
-      });
-
-      return pdfPath;
-   },
    removeFileType: (fileType) => ipcRenderer.invoke('removeFileType', fileType),
    addFileType: (fileType) => ipcRenderer.invoke('addFileType', fileType),
    getRemovedFileTypes: () => ipcRenderer.invoke("getRemovedFileTypes"),
